# ansible/deploy/files/api.yaml
openapi: 3.0.3
info:
  title: Salesforce SME DAT
  version: 1.0.0
x-amazon-apigateway-request-validators:
    all:
      validateRequestBody" : true,
      validateRequestParameters" : true
    params-only:
      validateRequestBody : false
      validateRequestParameters : true
x-amazon-apigateway-request-validator: params-only
paths:
  /index.html:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/index.html"
        responses:
          '200':
            statusCode: 200
          '304':
            statusCode: 304
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /css/salesforce-lightning-design-system.min.css:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/css/salesforce-lightning-design-system.min.css"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /css/app.css:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/css/app.css"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /config.json:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/config.json"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /js/chunk-vendors.js:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/js/chunk-vendors.js"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /js/app.js:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/js/app.js"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /assets/images/icon.png:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/assets/images/icon.png"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /assets/images/themes/oneSalesforce/banner-brand-default.png:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/assets/images/themes/oneSalesforce/banner-brand-default.png"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /assets/icons/utility-sprite/svg/symbols.svg:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/assets/icons/utility-sprite/svg/symbols.svg"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /fonts/SalesforceSans-Light.woff2:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/fonts/SalesforceSans-Light.woff2"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /fonts/SalesforceSans-Regular.woff2:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/fonts/SalesforceSans-Regular.woff2"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /fonts/SalesforceSans-Light.woff:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/fonts/SalesforceSans-Light.woff"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /fonts/SalesforceSans-Regular.woff:
    get:
      responses:
        '200':
          description: OK
        '304':
          description: Not Modified
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/fonts/SalesforceSans-Regular.woff"
        responses:
          "200":
            statusCode: "200"
          "304":
            statusCode: "304"
          default:
            statusCode: "500"
        httpMethod: "GET"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"                                
  /journey/execute:
    post:
      summary: Endpoint used to execute the journey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                inArguments:
                  type: array
                  items: {}
                outArguments:
                  type: array
                  items: {}
                activityObjectID:
                  type: string
                  example: 4ce1a85a-14d4-4216-8574-8db55028db13
                journeyId:
                  type: string
                  example: 075fe2eb-08d2-40d5-8fe9-0d0b80506226
                activityId:
                  type: string
                  example: 4ce1a85a-14d4-4216-8574-8db55028db13
                definitionInstanceId:
                  type: string
                  example: fc86adbe-864a-47b5-b867-93566a2aa9bf
                activityInstanceId:
                  type: string
                  example: 112c5bee-aa6f-4971-9f8c-b8bc6d7beb7b
                keyValue:
                  type: string
                  example: test@test.com
                mode:
                  type: integer
                  example: 0
      responses:
        '200':
          description: Response with id and resolved schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  schedule:
                    type: string
                    example: dddd, MMMM DD, YYYY hh:mm A
        '400':
          description: Validation error.
        '500':
          description: Unexpected error.
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/journey/execute"
        responses:
          "200":
            statusCode: "200"
          "400":
            statusCode: "400"
          default:
            statusCode: "500"
        httpMethod: "POST"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /journey/save:
    post:
      summary: Endpoint used to save the journey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                activityObjectID:
                  type: string
                  example: 53db1f41-3983-4f96-a45a-d7d8f968fb22
                interactionId:
                  type: string
                  example: 50ed1f0c-d74c-44de-99b0-9721ce97bbe1
                originalDefinitionId:
                  type: string
                  example: 075fe2eb-08d2-40d5-8fe9-0d0b80506226
                interactionKey:
                  type: string
                  example: bdfe0d45-f62b-8ef5-82c1-82cde48023dc
                interactionVersion:
                  type: string
                  example: '3'
      responses:
        '200':
          description: Response with status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '500':
          description: Unexpected error.
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/journey/save"
        responses:
          "200":
            statusCode: "200"
          default:
            statusCode: "500"
        httpMethod: "POST"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /journey/publish:
    post:
      summary: Endpoint used to publish the journey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                activityObjectID:
                  type: string
                  example: 53db1f41-3983-4f96-a45a-d7d8f968fb22
                interactionId:
                  type: string
                  example: 50ed1f0c-d74c-44de-99b0-9721ce97bbe1
                originalDefinitionId:
                  type: string
                  example: 075fe2eb-08d2-40d5-8fe9-0d0b80506226
                interactionKey:
                  type: string
                  example: bdfe0d45-f62b-8ef5-82c1-82cde48023dc
                interactionVersion:
                  type: string
                  example: '3'
                isPublished:
                  type: boolean
      responses:
        '200':
          description: Response with status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '500':
          description: Unexpected error.
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/journey/publish"
        responses:
          "200":
            statusCode: "200"
          default:
            statusCode: "500"
        httpMethod: "POST"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /journey/stop:
    post:
      summary: Endpoint used to stop the journey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                activityObjectID:
                  type: string
                  example: 53db1f41-3983-4f96-a45a-d7d8f968fb22
                interactionId:
                  type: string
                  example: 50ed1f0c-d74c-44de-99b0-9721ce97bbe1
                originalDefinitionId:
                  type: string
                  example: 075fe2eb-08d2-40d5-8fe9-0d0b80506226
                interactionKey:
                  type: string
                  example: bdfe0d45-f62b-8ef5-82c1-82cde48023dc
                interactionVersion:
                  type: string
                  example: '3'
      responses:
        '200':
          description: Response with status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '500':
          description: Unexpected error.
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/journey/stop"
        responses:
          "200":
            statusCode: "200"
          default:
            statusCode: "500"
        httpMethod: "POST"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /journey/validate:
    post:
      summary: Endpoint used to stop the journey
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                activityObjectID:
                  type: string
                  example: 53db1f41-3983-4f96-a45a-d7d8f968fb22
                interactionId:
                  type: string
                  example: 50ed1f0c-d74c-44de-99b0-9721ce97bbe1
                originalDefinitionId:
                  type: string
                  example: 075fe2eb-08d2-40d5-8fe9-0d0b80506226
                interactionKey:
                  type: string
                  example: bdfe0d45-f62b-8ef5-82c1-82cde48023dc
                interactionVersion:
                  type: string
                  example: '3'
      responses:
        '200':
          description: Response with status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '500':
          description: Unexpected error.
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/journey/validate"
        responses:
          "200":
            statusCode: "200"
          default:
            statusCode: "500"
        httpMethod: "POST"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"
  /token:
    post:
      summary: Endpoint used to validate token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  example: 53db1f41-3983-4f96-a45a-d7d8f968fb22
      responses:
        '200':
          description: Response with status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '500':
          description: Unexpected error.    
      x-amazon-apigateway-integration:
        type: "http_proxy"
        uri: "https://${stageVariables.APIENV_TARGET_ENDPOINT}/token"
        responses:
          "200":
            statusCode: "200"
          default:
            statusCode: "500"
        httpMethod: "POST"
        connectionType: "VPC_LINK"
        connectionId: "${stageVariables.APIENV_VPCLINK}"     